+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.0.5] - 2023-03-19
+
+### Added
+
+- Polish notation calculator example.
+- Benchmarks statistics.
+
 ## [0.0.4] - 2023-03-19
 
 ### Fixed